+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/domain/MyReviewDesign.java
@@ -16,8 +16,8 @@ public class MyReviewDesign {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "my_review_design_id")
-	private Long id;
+	private Integer id;
 
-	private Long reviewDesignId;
-	private Long shopAdminId;
+	private Integer reviewDesignId;
+	private Integer shopAdminId;
 }
